requirements.txt
celery==4.4.0rc3
dj-database-url==0.4.2
Django==3.1.3
django-redis==4.8.0
django-redis-cache==1.7.1
psycopg2-binary
redis==3.2.0
six==1.11.0
requests==2.25.0
django-extensions==3.0.9
djangorestframework==3.12.2
djangorestframework-composed-permissions==0.2.1
django-phonenumber-field==1.3.0
django-filter==2.4.0
factory-boy==3.1.0
mock==2.0.0
kombu==4.6.4
Pillow==8.0.1
django-polymorphic==2.1.2
djangorestframework-gis==0.14
djangorestframework-filters==0.11.1
gunicorn==19.9.0
jupyter
ipython[notebook]==7.19.0
django-cors-headers==2.2.0
djangorestframework-simplejwt==4.6.0
django-tqdm==0.0.3
werkzeug==1.0.1